:necktie: Feat: remove() 함수 테스트하기

+++ src/user/user.service.ts
@@ -42,7 +42,15 @@ export class UserService {
     return this.userRepository.findOne({ where: { id } });
   }
 
-  remove(id: number) {
-    return this.userRepository.delete({ id });
+  async remove(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException('존재하지 않는 사용자입니다!');
+    }
+
+    await this.userRepository.delete({ id });
+
+    return id;
   }
 }
